+++ babel.config.js
@@ -1,0 +1,12 @@
+module.exports = {
+    extends: '@krakenjs/grumbler-scripts/config/.babelrc-browser',
+    env: {
+        test: {
+            presets: [['@babel/env', { targets: { node: 'current' } }]]
+        }
+    },
+    plugins: [
+        ['@babel/plugin-proposal-private-methods', { loose: true }],
+        ['@babel/plugin-proposal-private-property-in-object', { loose: true }]
+    ]
+};