+++ src/library/controllers/modal/interface.js
@@ -1,5 +1,5 @@
 import arrayFind from 'core-js-pure/stable/array/find';
-import { ZalgoPromise } from 'zalgo-promise/src';
+import { ZalgoPromise } from '@krakenjs/zalgo-promise/src';
 
 import {
     logger,
@@ -30,8 +30,10 @@ const memoizedModal = memoizeOnProps(
         onReady,
         onCalculate,
         onApply,
+        onShow,
         onClose,
         channel,
+        cspNonce,
         integrationIdentifier
     }) => {
         addPerformanceMeasure(PERFORMANCE_MEASURE_KEYS.FIRST_MODAL_RENDER_DELAY);
@@ -48,8 +50,10 @@ const memoizedModal = memoizeOnProps(
             onReady,
             onCalculate,
             onApply,
+            onShow,
             onClose,
             channel,
+            cspNonce,
             integrationIdentifier
         });
         // Fired from inside the default onReady callback