+++ src/library/zoid/message/component.js
@@ -1,7 +1,8 @@
 import stringStartsWith from 'core-js-pure/stable/string/starts-with';
-import { ZalgoPromise } from 'zalgo-promise/src';
-import { uniqueID, getCurrentScriptUID } from 'belter/src';
-import { create } from 'zoid/src';
+import { SDK_SETTINGS } from '@paypal/sdk-constants/src';
+import { ZalgoPromise } from '@krakenjs/zalgo-promise/src';
+import { uniqueID, getCurrentScriptUID } from '@krakenjs/belter/src';
+import { create } from '@krakenjs/zoid/src';
 
 import {
     getMeta,
@@ -18,11 +19,14 @@ import {
     getOrCreateStorageID,
     getStageTag,
     getFeatures,
+    getNonce,
     ppDebug,
     isScriptBeingDestroyed,
+    getScriptAttributes,
     getDevTouchpoint,
     getMerchantConfig
 } from '../../../utils';
+
 import validate from './validation';
 import containerTemplate from './containerTemplate';
 
@@ -177,14 +181,13 @@ export default createGlobalVariableGetter('__paypal_credit_message__', () =>
                 type: 'function',
                 queryParam: false,
                 value: ({ props }) => {
-                    const { onHover } = props;
+                    const { index, onHover } = props;
                     let hasHovered = false;
 
                     return ({ meta }) => {
-                        const { index } = props;
-
                         if (!hasHovered) {
                             hasHovered = true;
+
                             logger.track({
                                 index,
                                 et: 'CLIENT_IMPRESSION',
@@ -203,9 +206,10 @@ export default createGlobalVariableGetter('__paypal_credit_message__', () =>
                 queryParam: false,
                 value: ({ props }) => {
                     const { onReady } = props;
-                    return ({ meta, activeTags, deviceID, requestDuration }) => {
-                        const { account, merchantId, index, modal, getContainer, messageRequestId } = props;
+                    return ({ meta, activeTags, deviceID, requestDuration, messageRequestId }) => {
+                        const { account, merchantId, index, modal, getContainer } = props;
                         const { trackingDetails, offerType, ppDebugId } = meta;
+                        const partnerClientId = merchantId && account.slice(10); // slice is to remove the characters 'client-id:' from account name
 
                         ppDebug(`Message Correlation ID: ${ppDebugId}`);
 
@@ -234,6 +238,7 @@ export default createGlobalVariableGetter('__paypal_credit_message__', () =>
                                     type: 'message',
                                     messageRequestId,
                                     account: merchantId || account,
+                                    partnerClientId,
                                     trackingDetails
                                 }
                             };
@@ -366,6 +371,12 @@ export default createGlobalVariableGetter('__paypal_credit_message__', () =>
                 value: getLibraryVersion,
                 debug: ppDebug(`Library Version: ${getLibraryVersion()}`)
             },
+            integrationType: {
+                type: 'string',
+                queryParam: true,
+                value: () => __MESSAGES__.__TARGET__,
+                debug: ppDebug(`Library Integration: ${__MESSAGES__.__TARGET__}`)
+            },
             deviceID: {
                 type: 'string',
                 queryParam: true,
@@ -416,6 +427,15 @@ export default createGlobalVariableGetter('__paypal_credit_message__', () =>
                 required: false,
                 value: getStageTag
             },
+            partnerAttributionId: {
+                type: 'string',
+                queryParam: true,
+                required: false,
+                value: () => (getScriptAttributes() ?? {})[SDK_SETTINGS.PARTNER_ATTRIBUTION_ID] ?? null,
+                debug: ppDebug(
+                    `Partner Attribution ID: ${(getScriptAttributes() ?? {})[SDK_SETTINGS.PARTNER_ATTRIBUTION_ID]}`
+                )
+            },
             devTouchpoint: {
                 type: 'boolean',
                 queryParam: true,
@@ -428,11 +448,11 @@ export default createGlobalVariableGetter('__paypal_credit_message__', () =>
                 required: false,
                 value: getFeatures
             },
-            integrationType: {
+            cspNonce: {
                 type: 'string',
-                queryParam: true,
                 required: false,
-                value: () => __MESSAGES__.__TARGET__
+                default: getNonce,
+                value: validate.cspNonce
             }
         }
     })