+++ tests/functional/v2/config/FR/index.js
@@ -1,0 +1,1 @@
+export { DEV_FR_SHORT_TERM } from './DEV_FR_SHORT_TERM';